fix(doctor): remove internal-only diagnostic checks from user-facing output
A few checks in nrv doctor, the end of nrv update, and nrv validate squad
were meant only for the owner's own release tooling, never for a regular
install. They ran unconditionally, so every user saw them and had no way
to act on what they reported. Removed from user-facing output; the
internal tooling that actually needs these checks moved to infrastructure
that isn't part of this repository.
