+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to the Nirvana-OS engine. Versions map to GitHub releases
 
 ## Unreleased
 
+### Less diagnostic noise in `nrv doctor`, `nrv update` and `nrv validate squad`
+
+A few checks in `nrv doctor`, the end of `nrv update`, and `nrv validate squad` were meant only for the owner's own release tooling, never for a regular install — but they ran unconditionally, so every user saw them and had no way to act on what they reported. Removed from user-facing output; the internal tooling that actually needs them moved to internal infrastructure that isn't part of this repository.
+
 ## 0.12.6 — 2026-08-31
 
 ### Glance: the org-chart tab is now editable, and two real bugs in it are fixed
