+++ docs/architecture/validate-gate.md
@@ -136,7 +136,7 @@ O schema em Zod, `MindCloneManifestSchema` (`skills/_shared/validators/validator
 
 O manifesto da biblioteca é são; o workflow não é. Medido em 26/08/2026 sobre 204 squads: 0 de 5.774 componentes declarados faltam e 705 de 705 `invoke.ref` resolvem, mas **160 de 1.740 `task:` e 180 de 2.786 `agent:` apontam para nada**, 56 passos carregam o prompt inline em vez de referenciar uma task, 15 workflows são órfãos, e o grafo aparece em oito dialetos dos quais só `steps[]` está na spec. O catálogo existe para nomear isso.
 
-**A severidade segue o protocolo do manifesto.** Sob `protocol: "6.0"` as regras de workflow são erro; sob `"5.0"` as mesmas regras são aviso. É a promessa de compatibilidade inteira deste corte: os 204 squads instalados mantêm o veredito que já têm, e um squad que optou pela v6 entra limpo. Três regras fogem disso de propósito: o teto de corpo e o workflow órfão são conselho sob qualquer protocolo (são fatos sobre autoria, não sobre contrato), e `distribution_artifacts` é sempre aviso porque uma cópia instalada de pack legitimamente carrega `PROVENANCE.json`, `LICENSE.txt` e watermark — reprovar por isso impediria o comprador de validar a própria biblioteca. A tabela declara a severidade da v6 (o estado-alvo); cada finding carrega a que de fato se aplica.
+**A severidade segue o protocolo do manifesto.** Sob `protocol: "6.0"` as regras de workflow são erro; sob `"5.0"` as mesmas regras são aviso. É a promessa de compatibilidade inteira deste corte: os 204 squads instalados mantêm o veredito que já têm, e um squad que optou pela v6 entra limpo. Três regras fogem disso de propósito: o teto de corpo e o workflow órfão são conselho sob qualquer protocolo (são fatos sobre autoria, não sobre contrato), e `distribution_artifacts` é sempre aviso porque uma cópia instalada de pack legitimamente carrega `PROVENANCE.json` e `LICENSE.txt` — reprovar por isso impediria o comprador de validar a própria biblioteca. A tabela declara a severidade da v6 (o estado-alvo); cada finding carrega a que de fato se aplica.
 
 | Critério | Severidade | Fixer |
 |---|---|---|
